feat:implement parse the enum from file.
